+++ Distributed_MatMul/include/initUtilities.h
@@ -1,8 +1,8 @@
 #pragma once
 #include <stdlib.h>
 
-void initOrder(double* matrix, uint myRank, uint N, uint workSize, uint NPEs);
+void initOrder(double* matrix, uint myRank, size_t N, size_t workSize, uint NPEs);
 
-void initID(double* matrix, uint myRank, uint N, uint workSize, uint NPEs);
+void initID(double* matrix, uint myRank, size_t N, size_t workSize, uint NPEs);
 
-void initRandom(double* matrix, ulong nElements);
+void initRandom(double* matrix, size_t nElements);
